fix ModalControllerにstoreメソッドを移す
